+++ packages/openauth/src/client.ts
@@ -86,6 +86,11 @@ export interface Tokens {
    * The refresh token.
    */
   refresh: string
+
+  /**
+   * The number of seconds until the access token expires.
+   */
+  expiresIn: number
 }
 
 interface ResponseLike {
@@ -657,6 +662,7 @@ export function createClient(input: ClientInput): Client {
         tokens: {
           access: json.access_token as string,
           refresh: json.refresh_token as string,
+          expiresIn: json.expires_in as number,
         },
       }
     },
@@ -699,6 +705,7 @@ export function createClient(input: ClientInput): Client {
         tokens: {
           access: json.access_token as string,
           refresh: json.refresh_token as string,
+          expiresIn: json.expires_in as number,
         },
       }
     },