+++ packages/openauth/src/provider/oidc.ts
@@ -23,7 +23,7 @@ import { WellKnown } from "../client.js"
 import { OauthError } from "../error.js"
 import { Provider } from "./provider.js"
 import { JWTPayload } from "hono/utils/jwt/types"
-import { getRelativeUrl } from "../util.js"
+import { getRelativeUrl, lazy } from "../util.js"
 
 export interface OidcConfig {
   /**
@@ -105,20 +105,24 @@ export function OidcProvider(
   const query = config.query || {}
   const scopes = config.scopes || []
 
-  const wk = fetch(config.issuer + "/.well-known/openid-configuration").then(
-    async (r) => {
-      if (!r.ok) throw new Error(await r.text())
-      return r.json() as Promise<WellKnown>
-    },
+  const wk = lazy(() =>
+    fetch(config.issuer + "/.well-known/openid-configuration").then(
+      async (r) => {
+        if (!r.ok) throw new Error(await r.text())
+        return r.json() as Promise<WellKnown>
+      },
+    ),
   )
 
-  const jwks = wk
-    .then((r) => r.jwks_uri)
-    .then(async (uri) => {
-      const r = await fetch(uri)
-      if (!r.ok) throw new Error(await r.text())
-      return createLocalJWKSet((await r.json()) as JSONWebKeySet)
-    })
+  const jwks = lazy(() =>
+    wk()
+      .then((r) => r.jwks_uri)
+      .then(async (uri) => {
+        const r = await fetch(uri)
+        if (!r.ok) throw new Error(await r.text())
+        return createLocalJWKSet((await r.json()) as JSONWebKeySet)
+      }),
+  )
 
   return {
     type: config.type || "oidc",
@@ -131,7 +135,7 @@ export function OidcProvider(
         }
         await ctx.set(c, "provider", 60 * 10, provider)
         const authorization = new URL(
-          await wk.then((r) => r.authorization_endpoint),
+          await wk().then((r) => r.authorization_endpoint),
         )
         authorization.searchParams.set("client_id", config.clientID)
         authorization.searchParams.set("response_type", "id_token")
@@ -159,7 +163,7 @@ export function OidcProvider(
         const idToken = body.get("id_token")
         if (!idToken)
           throw new OauthError("invalid_request", "Missing id_token")
-        const result = await jwtVerify(idToken.toString(), await jwks, {
+        const result = await jwtVerify(idToken.toString(), await jwks(), {
           audience: config.clientID,
         })
         if (result.payload.nonce !== provider.nonce) {